agent/requirements.txt
groq
supabase
python-dotenv
pyautogui
pygetwindow
pycaw
comtypes
psutil
pyperclip
